Revert "Photo Gallery Section in the Dorm Page"
